BREAKING CHANGE: Refactor DocxStamper class to new package
The DocxStamper class has been moved from org.wickedsource.docxstamper to pro.verron.officestamper.core as per the recent refactoring plan. The import statements across all files using this class have also been adjusted accordingly. The legacy code in examples regarding the use of DocxStamper inspired diagnostic methods has been removed.
