+++ engine/src/main/java/org/wickedsource/docxstamper/processor/BaseCommentProcessor.java
@@ -1,8 +1,8 @@
 package org.wickedsource.docxstamper.processor;
 
-import org.wickedsource.docxstamper.DocxStamper;
 import pro.verron.officestamper.api.AbstractCommentProcessor;
 import pro.verron.officestamper.api.ParagraphPlaceholderReplacer;
+import pro.verron.officestamper.core.DocxStamper;
 
 /**
  * Base class for comment processors. The current run and paragraph are set by the {@link DocxStamper} class.
