+++ examples/src/main/java/pro/verron/officestamper/Main.java
@@ -7,7 +7,6 @@ import java.nio.file.Files;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
-import static pro.verron.officestamper.Examples.legacyStampDiagnostic;
 import static pro.verron.officestamper.Examples.stampDiagnostic;
 
 public class Main {
@@ -16,10 +15,6 @@ public class Main {
 
     public static void main(String[] args)
             throws Exception {
-        // Use the way prior to docx-stamper 1.6.8
-        legacyStampDiagnostic(createOutStream("LegacyDiagnostic-"));
-
-        // Use the way from docx-stamper 1.6.8 and with office-stamper
         stampDiagnostic(createOutStream("Diagnostic-"));
     }
